+++ src/app/Balloon.App.Notification/Hook/Subscription.php
@@ -189,10 +189,12 @@ class Subscription extends AbstractHook
         }
 
         $body = preg_replace_callback('/(\{(([a-z]\.*)+)\})/', function ($match) use ($collection) {
-            return $collection->getAttribute($match[2]);
+            return '';
+            //return $collection->getAttribute($match[2]);
         }, $this->body);
         $subject = preg_replace_callback('/(\{(([a-z]\.*)+)\})/', function ($match) use ($collection) {
-            return $collection->getAttribute($match[2]);
+            return '';
+            //return $collection->getAttribute($match[2]);
         }, $this->subject);
 
         $this->notifier->notify($this->user, null, $subject, $body);