+++ src/lib/Api/v1/File.php
@@ -496,9 +496,9 @@ class File extends Node
                 }
 
                 $name = Helper::filter($name);
-                $result = $collection->addFile($name, $content, $attributes)->getId(true);
+                $result = $collection->addFile($name, $content, $attributes)->getId();
 
-                return (new Response())->setCode(201)->setBody($result);
+                return (new Response())->setCode(201)->setBody((string) $result);
             }
         } catch (ForbiddenException $e) {
             throw new Exception\Conflict(
@@ -522,9 +522,9 @@ class File extends Node
                         throw new Exception\InvalidArgument('name must be a valid string');
                     }
 
-                    $result = $parent->addFile($name, $content, $attributes)->getId(true);
+                    $result = $parent->addFile($name, $content, $attributes)->getId();
 
-                    return (new Response())->setCode(201)->setBody($result);
+                    return (new Response())->setCode(201)->setBody((string) $result);
                 } catch (Exception\NotFound $e) {
                     throw new Exception('parent collection '.$parent_path.' was not found');
                 }