+++ src/lib/Api/v1/Group.php
@@ -15,6 +15,7 @@ namespace Balloon\Api\v1;
 use Balloon\Exception;
 use Balloon\Filesystem\Acl\Exception\Forbidden as ForbiddenException;
 use Balloon\Server;
+use Balloon\Server\AttributeDecorator;
 use Micro\Http\Response;
 use MongoDB\BSON\ObjectId;
 
@@ -35,14 +36,23 @@ class Group
     protected $server;
 
     /**
+     * Attribute decorator.
+     *
+     * @var AttributeDecorator
+     */
+    protected $decorator;
+
+    /**
      * Initialize.
      *
      * @param Server $server
+     * @param AttributeDecorator
      */
-    public function __construct(Server $server)
+    public function __construct(Server $server, AttributeDecorator $decorator)
     {
         $this->group = $server->getIdentity();
         $this->server = $server;
+        $this->decorator = $decorator;
     }
 
     /**
@@ -197,7 +207,7 @@ class Group
      */
     public function getAttributes(?string $id = null, ?string $name = null, array $attributes = []): Response
     {
-        $result = $this->_getGroup($id, $name)->getAttribute($attributes);
+        $result = $this->decorator->decorate($this->_getGroup($id, $name), $attributes);
 
         return (new Response())->setCode(200)->setBody($result);
     }